# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Blossom Buddy - API Documentation",
        "description": "Blossom Buddy app, backend documentation",
        "version": "1.0.0"
    },
    "paths": {
        "/api/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Register a new user",
                "description": "Registers a new user and returns an authentication token.",
                "operationId": "8a56853624e025573120a09a4c75d468",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "$ref": "#/components/schemas/User"
                                    },
                                    {
                                        "properties": {
                                            "password": {
                                                "type": "string",
                                                "format": "password",
                                                "example": "password123"
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User registered successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "access_token": {
                                            "type": "string"
                                        },
                                        "token_type": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Login a user",
                "description": "Authenticates a user and returns an authentication token.",
                "operationId": "a3b306d14572d1f4bd6c064b3233e7b8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "john.doe@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User authenticated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "access_token": {
                                            "type": "string"
                                        },
                                        "token_type": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid credentials",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid credentials"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Logout user",
                "description": "Logs out the authenticated user",
                "operationId": "logout",
                "responses": {
                    "200": {
                        "description": "Successful logout"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/me": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Get authenticated user",
                "description": "Returns the authenticated user's information",
                "operationId": "me",
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/plants": {
            "get": {
                "tags": [
                    "Plants"
                ],
                "summary": "Get list of plants",
                "description": "Returns list of plants",
                "operationId": "getPlantsList",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Plant"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Plants"
                ],
                "summary": "Store a new plant",
                "operationId": "storePlant",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Plant"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Plant created successfully"
                    }
                }
            }
        },
        "/api/plants/{common_name}": {
            "get": {
                "tags": [
                    "Plants"
                ],
                "summary": "Get plant by common name",
                "description": "Returns a single plant",
                "operationId": "getPlantByCommonName",
                "parameters": [
                    {
                        "name": "common_name",
                        "in": "path",
                        "description": "The common name of the plant",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Plant"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Plant not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Plants"
                ],
                "summary": "Update an existing plant",
                "description": "Updates a plant by common_name",
                "operationId": "updatePlantByCommonName",
                "parameters": [
                    {
                        "name": "common_name",
                        "in": "path",
                        "description": "common_name of the plant to update",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Plant"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Plant"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Plant not found"
                    },
                    "422": {
                        "description": "Unprocessable Content"
                    }
                }
            }
        },
        "/api/plants/{id}": {
            "delete": {
                "tags": [
                    "Plants"
                ],
                "summary": "Delete an existing plant",
                "description": "Delete a plant by id if it exists",
                "operationId": "deletePlant",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of the plant to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Plant successfully deleted"
                    },
                    "404": {
                        "description": "Plant not found"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Plant": {
                "title": "Plant",
                "properties": {
                    "common_name": {
                        "description": "The common name of the plant",
                        "type": "string",
                        "example": "Aloe Vera"
                    },
                    "watering_general_benchmark": {
                        "description": "The general watering benchmark of the plant",
                        "properties": {
                            "value": {
                                "type": "string",
                                "example": "1"
                            },
                            "unit": {
                                "type": "string",
                                "example": "week"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "User": {
                "title": "User",
                "properties": {
                    "firstName": {
                        "description": "The first name of the user",
                        "type": "string",
                        "example": "John"
                    },
                    "lastName": {
                        "description": "The last name of the user",
                        "type": "string",
                        "example": "Doe"
                    },
                    "email": {
                        "description": "The email of the user",
                        "type": "string",
                        "example": "john.doe@example.com"
                    }
                },
                "type": "object"
            }
        },
        "examples": {
            "plant": {
                "summary": "Example of a Plant",
                "value": {
                    "common_name": "Aloe Vera",
                    "watering_general_benchmark": {
                        "value": "1",
                        "unit": "week"
                    }
                }
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "tags": [
        {
            "name": "Authentication",
            "description": "Authentication"
        },
        {
            "name": "Plants",
            "description": "Plants"
        }
    ],
    "security": [
        {
            "sanctum": []
        }
    ]
}